+++ app/core/gann/ann.py
@@ -1,5 +1,5 @@
 import numpy
-import app.activation_function as activation_function
+import app.core.common.activation_function as activation_function
 
 
 class ANN:
