+++ app/core/utils/dominant_color_analyzer.py
