+++ test/test_core/utils/test_color_analyzer_dataset.py
