+++ test/test_core/utils/test_gann_trainer.py
@@ -1,8 +1,8 @@
 import unittest
 import numpy
 from unittest.mock import patch, MagicMock
-from app.gann import GANN
-from app.gann_trainer import ScoreEvaluator
+from app.core.gann.gann import GANN
+from app.core.utils.gann_trainer import ScoreEvaluator
 from builder.builders import A
 
 
@@ -12,9 +12,9 @@ class TestGANNTrainer(unittest.TestCase):
         self.assertEqual(len(gann_trainer.population), 5)
         self.assertIsInstance(gann_trainer.population[0][0], GANN)
 
-    @patch('app.gann_trainer.GANNTrainer.reproduction')
-    @patch('app.gann_trainer.GANNTrainer.selection')
-    @patch('app.gann_trainer.GANNTrainer.evaluation')
+    @patch('app.core.utils.gann_trainer.GANNTrainer.reproduction')
+    @patch('app.core.utils.gann_trainer.GANNTrainer.selection')
+    @patch('app.core.utils.gann_trainer.GANNTrainer.evaluation')
     def test_next_generation(self, mock_evaluation, mock_selection, mock_reproduction):
         gann_trainer = A.GANNTrainer.build()
         gann_trainer.next_generation()
@@ -23,8 +23,8 @@ class TestGANNTrainer(unittest.TestCase):
         mock_selection.assert_called()
         mock_reproduction.assert_called()
 
-    @patch('app.gann_trainer.GANNTrainer._update_test_score')
-    @patch('app.gann_trainer.GANNTrainer._update_training_score')
+    @patch('app.core.utils.gann_trainer.GANNTrainer._update_test_score')
+    @patch('app.core.utils.gann_trainer.GANNTrainer._update_training_score')
     def test_evaluation(self, mock_training_score, mock_test_score):
         gann_trainer = A.GANNTrainer.build()
         gann_trainer.population = [2, 3, 4]
@@ -33,7 +33,7 @@ class TestGANNTrainer(unittest.TestCase):
         mock_training_score.assert_called_with(4)
         mock_test_score.assert_called_with(4)
 
-    @patch('app.gann_trainer.ScoreEvaluator.calc_training_score')
+    @patch('app.core.utils.gann_trainer.ScoreEvaluator.calc_training_score')
     def test_update_training_score(self, mock_calc_score):
         mock_calc_score.return_value = 4
 
@@ -44,7 +44,7 @@ class TestGANNTrainer(unittest.TestCase):
 
         self.assertEqual(4, mock_individual[1])
 
-    @patch('app.gann_trainer.ScoreEvaluator.calc_test_score')
+    @patch('app.core.utils.gann_trainer.ScoreEvaluator.calc_test_score')
     def test_update_test_score(self, mock_calc_score):
         mock_calc_score.return_value = 3
 
@@ -55,8 +55,8 @@ class TestGANNTrainer(unittest.TestCase):
 
         self.assertEqual(3, mock_individual[2])
 
-    @patch('app.gann_trainer.GANNTrainer._remove_unfit_individuals')
-    @patch('app.gann_trainer.GANNTrainer.sort_population_by_training_score')
+    @patch('app.core.utils.gann_trainer.GANNTrainer._remove_unfit_individuals')
+    @patch('app.core.utils.gann_trainer.GANNTrainer.sort_population_by_training_score')
     def test_selection(self, mock_sort, mock_remove):
         gann_trainer = A.GANNTrainer.build()
         gann_trainer.selection()
@@ -98,7 +98,7 @@ class TestScoreEvaluator(unittest.TestCase):
         expected = 1 + 2 + 1
         self.assertEqual(result, expected)
 
-    @patch('app.gann_trainer.ScoreEvaluator._calc_score')
+    @patch('app.core.utils.gann_trainer.ScoreEvaluator._calc_score')
     def test_calc_gann_score(self, mock_calc_score):
         mock_calc_score.return_value = -5
 
@@ -115,7 +115,7 @@ class TestScoreEvaluator(unittest.TestCase):
         mock_gann.forward.assert_called_with([-2, 1, 6])
         self.assertEqual(result, expected)
 
-    @patch('app.gann_trainer.ScoreEvaluator._calc_gann_score')
+    @patch('app.core.utils.gann_trainer.ScoreEvaluator._calc_gann_score')
     def test_calc_gann_dataset_score(self, mock_calc_gann_score):
         mock_calc_gann_score.return_value = 5
 
@@ -136,7 +136,7 @@ class TestDatasetScoring(unittest.TestCase):
         dataset.test_dataset = [2, 4]
         self.score_evaluator = ScoreEvaluator(dataset)
 
-    @patch('app.gann_trainer.ScoreEvaluator._calc_gann_dataset_score')
+    @patch('app.core.utils.gann_trainer.ScoreEvaluator._calc_gann_dataset_score')
     def test_calc_training_score(self, mock_score):
         mock_score.return_value = 3
 
@@ -146,7 +146,7 @@ class TestDatasetScoring(unittest.TestCase):
         mock_score.assert_called_with(mock_gann, [1, 3])
         self.assertEqual(3, result)
 
-    @patch('app.gann_trainer.ScoreEvaluator._calc_gann_dataset_score')
+    @patch('app.core.utils.gann_trainer.ScoreEvaluator._calc_gann_dataset_score')
     def test_calc_test_score(self, mock_score):
         mock_score.return_value = -4
